Add modified versions of YAGSL modules for CTRE Pro
In order to use modified versions of the CTRE connection modules from YAGSL,
needed to modify several files to deal with the parsing chain.  Not sure if
this is the best long-term solution, but I think it will work for now.

The next step will be to modify the ``YAGSL_AzRBSI/*_RBSI.java`` hardware
control files (CANCoder, Pigeon2, TalonFX) to allow access to the Phoenix Pro
features we want access to.

	modified:   src/main/java/frc/robot/RobotContainer.java
	modified:   src/main/java/frc/robot/subsystems/swervedrive/SwerveSubsystem.java
	new file:   src/main/java/frc/robot/util/YAGSL_AzRBSI/CANCoderSwerve_RBSI.java
	new file:   src/main/java/frc/robot/util/YAGSL_AzRBSI/DeviceJson_RBSI.java
	new file:   src/main/java/frc/robot/util/YAGSL_AzRBSI/Pigeon2Swerve_RBSI.java
	new file:   src/main/java/frc/robot/util/YAGSL_AzRBSI/README
	new file:   src/main/java/frc/robot/util/YAGSL_AzRBSI/SwerveDriveJson_RBSI.java
	new file:   src/main/java/frc/robot/util/YAGSL_AzRBSI/SwerveParser_RBSI.java
	new file:   src/main/java/frc/robot/util/YAGSL_AzRBSI/TalonFXSwerve_RBSI.java