Remove gyro manual reset function
Not sure if this is what's causing PathPlanner to behave oddly, but it's
something not in the AK2025 templates.

The other thing to try will be to remove the IMU accelerometer readout from
the Accelerometer subsystem, with the thought that the extra read may be
messing with the PathPlanner gyro settings.

	modified:   src/main/java/frc/robot/RobotContainer.java
	modified:   src/main/java/frc/robot/subsystems/drive/Drive.java
	modified:   src/main/java/frc/robot/subsystems/drive/GyroIO.java
	modified:   src/main/java/frc/robot/subsystems/drive/GyroIONavX.java
	modified:   src/main/java/frc/robot/subsystems/drive/GyroIOPigeon2.java